Use Runes framework built by Carthage
Instead of using Git Submodules to bring in Runes as a dependency, it's
actually a lot cleaner (and faster) to just use Carthage as a
development dependency and assume users will run `carthage bootstrap`
when setting up. This way, we can rely on neat things like Carthage's
support for pre-built binaries.