Move JSONValue.mapDecode to a global function
This was kind of bolted onto JSONValue without too much benefit. It was
also pretty poorly named. It makes more sense to me as a global
function, seeing as it's essentially already a workaround for the fact
that we can't create an extension on Array with a type constraint of
JSONDecodable.