Add tests for decoding native Dictionary and plists
Argo was made specifically for decoding JSON into model objects but the
implementation allows it to decode anything that can be cast to an
AnyObject and conforms to JSONDecodable into model objects. This means
native Swift types like Dictionary and Array as well as anything that
those types represent. For instance, plists are represented as native
dictionaries and arrays to they can be decoded into models by Argo. In
theory, any parser that parses data to a Swift object can be decoded
into models by Argo. XML, YAML, etc. should all be possible.