Implement a Decoded type for error messaging
This introduces a `Decoded` type which is similar to a `Result` type.
It has a success state and 2 failure states. One failure state occurs if
there is a type mismatch and the other is for a missing key in the JSON.
This custom type allows us to retain some information about why a
decoding has failed. It also allows us to abstract the return type that
the user wants.

We have defined a global `decode` function for the user to use. It can
return an optional type, ignoring the failure data, or return the
`Decoded` type if the user would like to retain the error information.
Other return types like `Result` could also easily be implemented in a
particular project.

The 2 different error states will function a bit differently. If the
user wants an optional value from the decoding, indicated by `<|?` or
`<||?` the `Decoded` type will allow failure only if the key is missing
and not if the type is a mismatch.