Add a function to convert JSON to .None if = .Null
With the addition of the `Decoded` type we decided to return a Type
Mismatch whenever the type asked for did not align with the `JSON`. In
the case of Null, we want to assume that it means that the key is just
missing and not of the wrong type. This adds code to convert the `JSON`
into an optional and specifically return `.None` when the `JSON` ==
`.Null`.