Add `<|>` alternative operator
This operator chooses between either the left hand side or the right
hand side, whichever contains a successful value. If both sides are
empty, the right hand side will be returned.

Usage is as follows:
<*> j <| "possible-key" <|> j <| "possible-key2"

The precedence is set a bit below the precedence of the other operators
(i.e. `<|`) so that parentheses are not necessary when constructing
expressions.