Add nil coalescing operator. The use case for this operator is any json result where a value can be in multiple places due to historical API cruft. One example is the Instagram API, where the outer data envelope can look either like this (simplified): {"meta": {"code": 200}...} or like this: {"code": 200...} The second version only happens for specific error responses
Use it as follows:
<*> (j <|? ["meta", "code"]) ?? (j <| "code")

The operator requires the left hand side to be of the type Decoded<T?> (i.e. use <|?) otherwise it cannot safely work if the left hand side is nil. Also, one of the two sides needs to be non-nil, so that  the expression evaluates to T and not T?.