Use generic value type extension
The allows us to implement functions on generic value types like Array
where `T` is `Decodable`. We can implement the `decode` function for
Array, Dictionary, and Optional. This also removes all the operators
from Argo except 1. They were initially put in place to help the
compiler with type inference. It seems that they are not needed anymore.
We should look at a more complex project before merging this into
master.