Add more curry functions and use static inits
We can now use static init funtions on structs to curry an initializer
instead of writing `create` functions. However, it seems that the
"expression is too complex" compilation error happens on very simple
models. If this doesn't get better, the `create` function might still be
the best option for most cases.