This commit introduces extensions to Decodable providing default implementations of `decode` for certain `RawRepresentable` types.
For instance, an enum backed by a String can now be implicitly decoded by Argo simply by declaring that it is `Decodable`.

An implementation using `NSNumber`, which is the most generally applicable type, is introduced too. We might want to copy paste a ton more if we like this approach.