Add tests for decoding dictionaries
Turns out, we weren't testing the decoding of dictionaries anywhere.
That seems wrong, so this adds those tests and (thankfully) proves that
our current implementation works as expected.

Also, we ran into a swift bug that I actually hit the other day as well
that I need to file a radar on. The title of which will probably end up
being:

> flatMapping custom monadic type with static function defined in a
> protocol extension results in compiler crash

That's why we're duplicating the meat of `flatMap` for the purpose of
the test.