Manually set bitcode modes
A framework with bitcode enabled requires the following flag to create
and fill the bitcode section, `-fembed-bitcode`. However, this flag is
available only in archive builds, not in normal builds where it's used
the flag `-fembed-bitcode-marker` which creates a bitcode section but
has a size of 1.

Carthage doesn't use archive action to build frameworks which means the
resulting framework will contain a bitcode section but not properly
filled as expected, leading to some issues in the submission of apps
to TestFlight for example.