Use `T` instead of `A` for generics in operators
This is a holdover from when we thought Swift was going to use Haskell
style generics (`A`, `B`, `C`, etc) instead of using named generic
types. We still don't really have a good name for these types, but the
swift community has essentially standardized on `T` for unnamed generic
types, so we should use the same.