Change Argo to lazy parse json
To lazily parse json we change the `JSON` type to not be recursive.
Arrays and Objects no longer parse their values when the json is parsed
into a `JSON` value. Instead, we call parse on the values just before
they are decoded. The major pitfall here is that there is duplicate
parsing since previously parsed json will have to be parsed again if
used again.

Looking at the performance tests: Before this change, parsing took about
137ms and decoding took 37ms totalling 174ms. After this change, parsing
took 0ms (this makes sense) and decoding took 100ms totalling 100ms. So
we see that decoding took longer since we moved the parsing into it, but
we see the overall time is shorter and even shorter than just parsing
was.