Add a Boolean case to JSON
Previously, we were overloading `JSON.Number` to include boolean values.
This was done because we believed that there was no way to determine if
an `NSNumber` returned from `NSJSONSerialization` was actually a boolean
value or not. Turns out, this was incorrect.

Under the covers, `NSNumber` is toll-free bridged with `CFNumberRef`.
And `CFNumberRef` has a specific implementation for boolean values
(`CFBooleanRef`). So we can actually get the type ID for `CFBooleanRef`
and just check it against the type ID for `self` and if that returns
`true`, then we have a boolean value and can handle it accordingly.