Use a class with a lazy variable for lazy decoding
We have to use a class here since lazy causes mutability and that
doesn't work well with structs. I think it even my be impossible the way
we've coded Argo. Anyhow, using the class in this manor reduces the
total decoding time by 60% (0.06s) of the first commit in this PR. I
really dislike the mutability of it all.