Remove most uses of reduce.
Argo’s performance is quadratic over the size of the collection being
decoded, due to the use of reduce[1]. This commit removes most usages of
reduce in exchange for a for loop of the same functionality. This
improves the performance of decoding an array of over 9000 simple models
on an iPhone 4S by ~2600% (76.4s to 2.9s). See the conversation in the
performance GitHub issue[2] for my analysis and evolution of this code.

[1]: http://airspeedvelocity.net/2015/08/03/arrays-linked-lists-and-performance/
[2]: https://github.com/thoughtbot/Argo/issues/295