Fix performance test for big_data.json
At some point, the big_data.json fixture became out of sync with the
TestModel type, and so decoding was failing, making these tests useless.
I've generated a new big_data.json fixture that matches the expected
TestModel properties, as well as increasing the size to be able to
really measure performance issues. I've also added a test to ensure that
decoding completed successfully so that we don't run into this issue in
the future.