Allow boolean values to be represented by numbers
Servers commonly represent boolean values as numbers even though this is
clearly wrong and they should be sending `true`/`false`.

Anyway, we can make it easier to deal with this by allowing numbers to
represent boolean values as well. This essentially fixes compatibility
with earlier versions of Argo, which did this anyway.