Add `JSON.asObject` property
This provides an `asObject` property that converts a `JSON` type to an
`AnyObject`. It's the inverse of the existing `JSON.init(_ json: AnyObject)`.

One cool thing, probably not that useful of thing, this allows you to do is set
up a `JSON` structure and convert it to a JSON string. That might look like:

```swift
let swiftJSON: JSON = .array([
    .object([
      "string": .string("hello world"),
      "stringArray": .array([
        .string("hello"),
        .string("Again")
      ])
    ])
])

do {
  let data = try JSONSerialization.data(withJSONObject: swiftJSON.asObject, options: [])
    let jsonString = String(data: data, encoding: .utf8)
    print(jsonString)
} catch {
  print("failed json encoding")
}
```

I'm not sure how many real world things this applies to, but it seems like a
generally useful utility to expose.

The example where I wanted this, was when attempting to implement
[DataRepresentable] on Argo's `JSON` type. It expects me to be able to convert
the object to `NSData` so it can be persisted to disk so this came in handy. We
could call that something like "serialization" or "encoding".

[DataRepresentable]: https://github.com/Haneke/HanekeSwift/blob/master/Haneke/Data.swift#L18-L21

With that in mind, it seems like this might be useful for someone who does want
to implement some sort of persistance, as this gives them a way to convert this
type to something convenient to persist. It doesn't solve encoding an actual
model object which would be nice, but perhaps this could lead to that at some
point.