Reveal Runes as a hard dependency
A while back, we used git submodules and linking trickery to hide the fact
that Argo uses Runes as a dependency for declaring monadic operator
definitions. This seemed like a good idea at the time, and has _mostly_ worked
well, but for Swift Package Manager we can't actually use the same technique
to link in Runes.swift

Additionally, using submodules in this way actually caused some subtle issues
for users:

 - If CocoaPods users wanted to specify a git branch or tag or commit or
   anything, they would also need to pass the `:submodules => true` flag in
   their podfile. This made us unlike basically any dependency.

 - Git Submodules subtly changed their behavior recently, which actually
   _completely broke_ Argo for CocoaPods users on git 2.9. It was later fixed,
   but we had to discover/debug/fix the problem ourselves, since we were
   apparently the only people that hit it. This is partially due to the fact
   that I aggressively update tools like git, but it's also a sign that we're
   a bit of an edge case, and that's not a super comfortable place to be. #
   Please enter the commit message for your changes. Lines starting

By declaring Runes as a hard, external dependency, we can get around these
issues, simplify our project setup, and move ourself to be in a better place
for supporting Swift Package Manager through the same mechanism.

The downside to this is that we've moved back to needing users to manually
import/link Runes into their files/apps when using Argo. I wish this wasn't
the case, but it still feels like the most reasonable way forward.