Move uint decoding back into the numerical type
This was originally broken out because of wanting to work around the compiler
complexity issues. However, we can actually get around this on our own by
breaking the decode operation into two distinct steps by assigning the
partially applied function to a local variable and then continuing. Since we
can do that, it makes more sense to move this all back under the main
Numerical type.