Add Carthage directories to framework search paths
When building with Carthage locally, Argo uses the local checkouts of the
dependencies to build and link against. However, when using Argo as a
dependency, those checkouts don't exist. This means that when Argo tries to
build, it can't find Runes and so the build fails.

Unfortunately, if we move to a more carthage centric workflow where we use the
built binaries to link against, that ends up breaking the workflow for users
that want to use Git Submodules for their dependency management (because they
won't have those frameworks in those locations).

We can work around this while still allowing users to use submodules by adding
the build directories to the framework search paths _without_ explicitly using
those built frameworks in our project. This means that Carthage users will be
able to see the Runes dependency, and submodules users will be able to use the
project and implicit dependencies.

The only downside here is that this might result in warnings for submodules
users. I think that's a reasonable trade-off.