Update cast method for Float/UInt
With Swift 3.0.1, Swift changed some things about the way it casts to
NSNumber[1]. As a result (I think?) using `as` to cast `UInt` and `Float`
values causes an EXC_BAD_ACCESS crash when decoding Swift dictionaries.
There are two solutions to this issue:

1. Explicitly cast these values to `NSNumber` in the test data
2. Use `.uintValue` and `.floatValue` to convert from `NSNumber` to the
expected type.

Option 1 seemed like a poor choice because it meant users of Argo would
potentially need to make source changes. This would have been
unfortunate.

Therefore, we choose to go with option 2, which should fix the issue for
the users relying on this functionality, be backwards compatible with
Swift 3.0, and be invisible to the users that have never used Argo to
parse `UInt` values from a Swift dictionary (most likely the entirety of
the Argo codebase).

[1]: https://github.com/apple/swift-evolution/blob/master/proposals/0139-bridge-nsnumber-and-nsvalue.md