Remove remaining NSNumber coercions
There were two other places that we were using `as` coercion to move
from NSNumber to our expected types. These actually don't compile at all
on Linux, which is interesting, so it's probably best if we just remove
these.

Also taking this opportunity to clean up the style on our
RawRepresentable implementations as well.