Rename Decode operator file for SPM
Swift Package Manager has a hard time when files are named the same.
This deficiency is usually caught when building the lib, but it looks
like `swift build` isn't using a case insensitive comparison when
checking for conflicting filenames. The end result for users is that our
lib _builds_ fine, but when people try to use it they get errors about
ambiguous function use.

The simplest fix on our end is to rename one of our files. It doesn't
_really_ matter which one we rename, so I just picked one that I felt I
could rename logically easily.