Disambiguate Swift.Decodable and Argo.Decodable
The stdlib now includes a built-in Decodable protocol to represent types
that can be decoded using the new Codable stuff. This means that our
conformance to `Decodable` is ambiguous. This is easily solved by
disambiguating with the `Argo` namespace.

Ultimately, we should look at renaming this entirely for our next major
release, but this at least fixes the build.