Use a file-based lookup method for fixtures
When running our tests via `swift-test`, we can't use the bundle stuff
to find out test fixtures. This means that we need to move to a
file-based approach.

The simplest solution is to use a relative path lookup based on where
the filetype-specific loader lives. That way, we can look up local
relative files quickly and easily.