Remove test support from Xcode
Even before Swift 4 and our ability to use `swift-test` reasonably, we
_technically_ required developers to run 3 different test suites _and_
build the project to ensure that everything was wired up correctly. This
was already pretty unmaintainable, and most people didn't actually do it
(myself included). Instead, we relied on CI to run these tests for us.

With Swift 4 and better support for `swift-test`, we would be adding yet
_another_ source of truth to that stack. To make things worse, it's
_incredibly_ easy to get tests to pass on one suite but not on the other
(see the changes we needed to make for fixture lookup as an example). To
avoid even more overhead and confusion here I'd rather start skating to
where the puck is headed.

My assumption is that we'll eventually be dropping (explicit) support
for both CocoaPods _and_ Carthage, instead relying on Swift Package
Manager. This would mean that `swift-test` will be the ultimate source
of truth in the future, so we might as well make it our source of truth
now.

To do this, I'm going to start by simply removing support for running
the tests from the default Xcode project, as well as removing the
`bin/test` helper script. We'll re-add CI support for running our tests
down the line.