Split test fixtures out into a new module
Breaking this out into its own module is really just cosmetic, but it
makes it easier to control the public interface of this thing, and gives
me a chance to normalize the API between the json and plist readers.

This is a good example of something that was super easy with SPM but
would have been a nightmare if we were still supporting tests in Xcode.
We would have had to make a Fixtures helper-framework for every platform
we were testing on. Blech.