Remove <|| family of operators
We can now use conditional conformance to remove the need for the `<||` family
of operators, and instead make `[T]` conform directly to `Decodable` if `T:
Decodable`.

Note that this means we can _also_ do this for `[String: T: Decodable]`, which
means you can now use `<|` to decode directly to `[String: T]`.