Replace <| operator with generic subscript
The `<|` operator was necessary when we originally implemented Argo, but Swift
has new capabilities now that reduce the need to use a custom operator.
Additionally, the operator spelling itself actually conflicts with prior-art
that we weren't aware of when we added it. Specifically, `<|` is used in other
languages (and sometimes in Swift) as a "pipe backwards" operator, with the
type `(T -> U) -> T -> U`.

Given these concerns, as well as the additional complexity around fine-tuning
the precedence so that it works in an obvious way, I think it makes sense to
remove the operator in favor of a generic subscript (which only just became
available in Swift 4).